requirements.txt
requests
psutil
python-telegram-bot==12.6.1 
google-api-python-client>=1.7.11,<1.7.20
google-auth-httplib2>=0.0.3,<0.1.0
google-auth-oauthlib>=0.4.1,<0.10.0
aria2p>=0.9.0,<0.15.0
python-dotenv>=0.10
tenacity>=6.0.0
python-magic
beautifulsoup4>=4.8.2,<4.8.10
Pyrogram>=0.16.0,<0.16.10
TgCrypto>=1.1.1,<1.1.10
youtube_dl
megasdkrestclient>=0.1.1,<1.0.0
#python-magic-bin==0.4.14
lxml
xdcc
